requirements.txt
langchain
streamlit
openai
fpdf2
yagmail